feat: split common file to utility files
