+++ packages/server/src/util/error.ts
@@ -1,5 +1,5 @@
 import type { APIGatewayProxyResult } from 'aws-lambda';
-import { createResponse } from './common';
+import { createResponse } from '../common';
 
 export class ResponsibleError extends Error {
 	additionalInfo: Record<string, unknown>;
@@ -24,6 +24,17 @@ export class UnauthorizedError extends ResponsibleError {
 		return createResponse(401, {
 			success: false,
 			error: 401,
+			error_description: this.message,
+			...this.additionalInfo
+		});
+	}
+}
+
+export class NotFoundError extends ResponsibleError {
+	response(): APIGatewayProxyResult {
+		return createResponse(404, {
+			success: false,
+			error: 404,
 			error_description: this.message,
 			...this.additionalInfo
 		});
